initialized motors in closed loop control mode on setup. formatting, other minor changes

+++ src/uirover_hardware/include/uirover_hardware/uirover_canbus_interface.hpp
@@ -59,7 +59,7 @@ struct Odrive {
 };
 struct Wheel {
   std::string joint_name;
-  u_int32_t node_id;
+  u_int8_t node_id;
   double velocity_command;
   double position_state;
   double velocity_state;
